Agent execution could get stuck.
For some reason, when the ThreadPool is instantiated with more than one thread,
an agents  DeliberationRunnable could be scheduled for execution while another
instance of that runnable is still in the waiting queue (or perhaps still being
executed). What I assume is happening is that the same instance of such a
runnable is accepted by more than one thread at the time, resulting in a GridLock.

I have not yet been able to figure out
1. How exactly the same instances can be scheduled more than once, or
2. What exactly causes the grid lock, even when two instances are being scheduled at
	the same time.

The patch is to do a check if an instance of the runnable is already scheduled. Its not
ideal, but its the best sollution I have so far.